In .tcpalette docs: fix JSON mistakes, conform with BCP 47 (en-GB as opposed to en_GB), and change colour values to be strictly CSS valid, which should be the **only** valid way.
The way I see it, defining file formats is not an area to try and get
clever in. Even if a neater standard is possible, one must resist. The
world isn't neat and in its own special way, designing your software to
fit the world is the most fulfilling way to create - both to you and
others.

Make your format guessable, but also don't be so flexible that the
behaviour ends up seeming unpredictable or that your implementation
grows complicated.

After someone's used your interface, they should almost know how to
implement it themselves. There shouldn't be a ton of extra crappy
optional ways to do things,

unless you want to, i guess.